requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==6.0.1
flask-migrate==4.1.0
pymysql==1.1.1
python-dotenv==1.1.1
marshmallow==3.23.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
flask-socketio==5.4.1
python-socketio==5.11.4
flask-jwt-extended==4.6.0
flask-mail==0.10.0
werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
bcrypt==4.1.2 
gunicorn